unify handling of bodies in extern specs and check more places
