+++ prusti-contracts/prusti-specs/src/extern_spec_rewriter/impls.rs
@@ -92,12 +92,7 @@ fn rewrite_plain_impl(impl_item: &mut syn::ItemImpl, new_ty: Box<syn::Type>) -> 
                 ));
             }
             syn::ImplItem::Method(method) => {
-                if !is_stub(&method.block) {
-                    return Err(syn::Error::new(
-                        method.span(),
-                        "Unexpected method body. (Extern specs only define specifications.)",
-                    ));
-                }
+                check_is_stub(&method.block)?;
 
                 let (rewritten_method, spec_items) = generate_extern_spec_method_stub(
                     method,
@@ -131,21 +126,6 @@ fn rewrite_plain_impl(impl_item: &mut syn::ItemImpl, new_ty: Box<syn::Type>) -> 
     Ok(())
 }
 
-/// Recognizes method stubs, e.g. `fn foo();`.
-///
-/// The absence of a body is represented in a roundabout way:
-/// They have a body comprising a single verbatim item containing a single semicolon token.
-fn is_stub(block: &syn::Block) -> bool {
-    if block.stmts.len() != 1 {
-        return false;
-    }
-    if let syn::Stmt::Item(syn::Item::Verbatim(tokens)) = &block.stmts[0] {
-        return tokens.to_string() == ";";
-    } else {
-        return false;
-    }
-}
-
 fn rewrite_trait_impl(
     impl_item: syn::ItemImpl,
     new_ty: Box<syn::Type>,
@@ -172,6 +152,8 @@ fn rewrite_trait_impl(
                 ));
             }
             syn::ImplItem::Method(method) => {
+                check_is_stub(&method.block)?;
+                
                 let (rewritten_method, spec_items) = generate_extern_spec_method_stub(
                     &method,
                     &item_ty,
